requirements.txt
# requirements.txt
#pip install -r requirements.txt
python-telegram-bot==21.1.1
asyncpg
python-dotenv==1.0.1
aiohttp==3.13.3
pydantic_settings
openai
groq
PyPDF2
langchain
langchain-text-splitters
pgvector
python-docx